iPLAN/requirements.txt
numpy~=1.21.5
seaborn~=0.12.2
pyglet~=1.5.27
six~=1.16.0
gym~=0.26.2
torch~=1.13.1+cu116
tensorboard~=2.11.0
pyyaml~=6.0
sacred~=0.8.2
pillow~=9.3.0
pygame~=2.1.2
pandas~=1.3.5
scipy~=1.7.3
matplotlib~=3.5.3
setuptools~=65.5.0
pysc2~=4.0.0
enum34~=1.1.10
s2clientprotocol~=5.0.10.89165.0
click~=8.1.3